Add storing keys to redis table
