+++ spider/pyproject.toml
@@ -7,6 +7,8 @@ requires-python = ">=3.13"
 dependencies = [
     "bs4>=0.0.2",
     "celery[redis]>=5.5.3",
+    "hiredis>=3.2.1",
+    "redis>=5.2.1",
 ]
 
 [dependency-groups]
